Docker - Face extracting/requirements.txt
fastapi
uvicorn
opencv-python-headless
mediapipe
httpx
numpy
python-dotenv
python-multipart